+++ 04-binary-data/lab-04/lib/copyModule.js
@@ -1,0 +1,7 @@
+'use strict';
+
+const copyPaste = (bmp) => {
+  console.log('Copy Pasta! Hot and ready!.');
+};
+
+module.exports = copyPaste;
