+++ 04-binary-data/lab-04/lib/overwriteModule.js
@@ -1,0 +1,8 @@
+'use strict';
+
+const overWrite = (bmp) => {
+  console.log('The pen is mighty!');
+  bmp.buffer = Buffer.from('You\'re file has been overwritten!');
+};
+
+module.exports = overWrite;
