+++ 04-binary-data/lab-04/newIndex.js
@@ -1,9 +1,9 @@
 'use strict';
 
-// console.log('inside the filez');
-// console.log(process.argv);
-
 const fs = require('fs');
+const copyModule = require('./lib/copyModule');
+const overcastModule = require('./lib/overcastModule');
+const overwriteModule = require('./lib/overwriteModule');
 
 class Bitmap{
   constructor(filePath) {
@@ -24,32 +24,11 @@ class Bitmap{
   }
 }
 
-const copyPaste = (bmp) => {
-  console.log('Copy Pasta! Hot and ready!.');
-};
-
-const overCast = (bmp) => {
-  console.log('This is how you look on a typical Seattle day.');
-  for(let i = 122; i < 1146; i += 4) {
-    let avg = (bmp.buffer[i] + bmp.buffer[i + 1] + bmp.buffer[i + 2]) / 3;
-    bmp.buffer[i] = avg;
-    bmp.buffer[i +1] = avg;
-    bmp.buffer[i +2] = avg;
-  }
-};
-
-const overWrite = (bmp) => {
-  console.log('The pen is mighty!');
-  // bmp.buffer.write('I\'m overwriting your file.');
-  bmp.buffer = Buffer.from('You\'re file has been overwritten!');
-};
-
 //an object containing different methods for transformation
 const transformDictionary = {
-  //TODO: Make a greyscale transformation function in the future
-  copy: copyPaste,
-  write: overWrite,
-  seattle: overCast,
+  copy: copyModule,
+  write: overwriteModule,
+  seattle: overcastModule,
 };
 
 function readFileAndTransform() {
